requirements/requirements.txt
django==3.2.6
djangorestframework==3.12.4
psycopg2-binary==2.9.1
python-dotenv==0.19.0
django-cors-headers==3.8.0
requests
mock